+++ examples/python/Menus.py
@@ -198,6 +198,7 @@ class Review(ftk.Menu):
         self.addAction(actions.actions["Open"])
         self.addAction(actions.actions["Save"])
         self.addAction(actions.actions["SaveAs"])
+        self.addAction(actions.actions["ExportMarkers"])
         self.addAction(actions.actions["Close"])
         self.recentMenu = self.addSubMenu("Recent")
         self.addDivider();